240107 이상화 커밋 1. int -> long으로 전부 변경 2. 엔티티연관 관계 수정

+++ bootProject/src/main/java/com/busanit501/bootproject/dto/MatchingRoomDTO.java
@@ -1,8 +1,6 @@
 package com.busanit501.bootproject.dto;
 
 import com.busanit501.bootproject.domain.RoomStatus;
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -16,12 +14,12 @@ import java.sql.Timestamp;
 @Builder
 public class MatchingRoomDTO {
 
-    private int roomId;
-    private int hostId;
+    private long roomId;
+    private long hostId;
     private String title;
     private String description;
-    private int maxParticipants;
-    private int currentParticipants;
+    private long maxParticipants;
+    private long currentParticipants;
     private RoomStatus status;
     private Timestamp createdAt;
 }
