+++ bootProject/src/main/java/com/busanit501/bootproject/service/MatchingRoomService.java
@@ -7,14 +7,14 @@ import java.util.List;
 
 public interface MatchingRoomService {
     //매칭룸생성()
-    int addMatchingRoom(MatchingRoomDTO matchingRoomDTO, RoomParticipantsDTO roomParticipantsDTO);
+    long addMatchingRoom(MatchingRoomDTO matchingRoomDTO, RoomParticipantsDTO roomParticipantsDTO);
     //매칭룸업데이트
     void updateMatchingRoom(MatchingRoomDTO matchingRoomDTO);
     void exitMatchingRoom(MatchingRoomDTO matchingRoomDTO);
     void inviteMatchingRoom(MatchingRoomDTO matchingRoomDTO);
     //매칭룸삭세
-    void deleteMatchingRoom(int roomId);
-    void deleteRoomParticipants(int roomId,int userId);
+    void deleteMatchingRoom(long roomId);
+    void deleteRoomParticipants(long roomId,long userId);
     //매칭룸전체조회
-    List<MatchingRoomDTO> searchAllMatchingRoom(String keyword,int userId);
+    List<MatchingRoomDTO> searchAllMatchingRoom(String keyword,long userId);
 }
